pyproject.toml
[tool.poetry]
name = "test_raid"
version = "0.1.0"
description = ""
authors = ["Yaroslav Galiulin <yaroslav66@list.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
